# Conflicts:
#	src/main/java/com/around/aroundcore/config/AroundConfig.java
#	src/main/java/com/around/aroundcore/database/models/GameUser.java
#	src/main/java/com/around/aroundcore/database/services/ImageService.java
#	src/main/java/com/around/aroundcore/web/controllers/rest/auth/OAuthController.java
#	src/main/java/com/around/aroundcore/web/exceptions/api/image/ImageNullException.java
#	src/main/java/com/around/aroundcore/web/mappers/GameUserDTOMapper.java
#	src/main/java/com/around/aroundcore/web/services/apis/oauth/GoogleOAuthService.java

+++ src/main/java/com/around/aroundcore/database/services/ImageService.java
@@ -1,11 +1,16 @@
-package com.around.aroundcore.web.services;
+package com.around.aroundcore.database.services;
 
-import com.around.aroundcore.web.exceptions.api.image.*;
+import com.around.aroundcore.config.AroundConfig;
+import com.around.aroundcore.database.models.Image;
+import com.around.aroundcore.database.repositories.ImageRepository;
+import com.around.aroundcore.web.exceptions.image.*;
 import com.around.aroundcore.web.image.ImageType;
 import com.around.aroundcore.web.image.MultipartImage;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.io.FileUtils;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -17,11 +22,14 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.UUID;
 
 @Slf4j
 @Service
+@RequiredArgsConstructor
 public class ImageService {
+    private final ImageRepository imageRepository;
+    @Value("${around.home}")
+    private String aroundHome;
     @Value("${around.image.limit}")
     Integer imageLimitSize;
     @Value("${around.image.avatars.directory}")
@@ -31,14 +39,13 @@ public class ImageService {
     @Value("${around.image.default.directory}")
     String imageDefaultDirectory;
 
-    public String saveImage(MultipartFile imageFile, ImageType imageType) throws ImageSaveException, ImageSizeException, ImageEmptyException {
+    public Image createImage(MultipartFile imageFile, ImageType imageType) throws ImageSaveException, ImageSizeException, ImageEmptyException {
         if(imageFile.getSize() > FileUtils.ONE_MB*imageLimitSize){
             throw new ImageSizeException();
         }
         if(imageFile.isEmpty()){
             throw new ImageEmptyException();
         }
-        log.info(imageFile.getContentType());
         if(!imageFile.getContentType().equals("image/jpeg") && !imageFile.getContentType().equals("image/png")){
             throw new ImageTypeException();
         }
@@ -48,37 +55,51 @@ public class ImageService {
             throw new ImageConvertException();
         }
 
-        String uniqueFileName = UUID.randomUUID() + "-" + imageFile.getOriginalFilename();
+        Image image = Image.builder().build();
+        log.info(image.getUuid().toString());
+        String uniqueFileName = image.getUuid().toString()+".jpg";
         Path uploadPath;
+        String url;
         switch (imageType){
-            case AVATAR -> uploadPath = Path.of(imageAvatarDirectory);
-            case ICON -> uploadPath = Path.of(imageIconDirectory);
-            default -> uploadPath = Path.of(imageDefaultDirectory);
+            case AVATAR -> {
+                uploadPath = Path.of(imageAvatarDirectory);
+                url = aroundHome+ AroundConfig.URL_AVATAR+uniqueFileName;
+            }
+            case ICON -> {
+                uploadPath = Path.of(imageIconDirectory);
+                url = aroundHome+ AroundConfig.URL_ICON+uniqueFileName;
+            }
+            default -> {
+                uploadPath = Path.of(imageDefaultDirectory);
+                url = aroundHome+ AroundConfig.URL_IMAGE+uniqueFileName;
+            }
         }
         Path filePath = uploadPath.resolve(uniqueFileName);
-
         try{
             if (!Files.exists(uploadPath)) {
                 Files.createDirectories(uploadPath);
             }
             imageFile.transferTo(filePath);
-        } catch (IOException e) {
+            image.setUrl(url);
+            image.setFile(uniqueFileName);
+            imageRepository.save(image);
+        } catch (Exception e) {
             log.error(e.getMessage());
-            e.printStackTrace();
             throw new ImageSaveException();
         }
 
-        return uniqueFileName;
+        return image;
     }
-
+    public void create(Image image){
+        imageRepository.save(image);
+    }
     public byte[] loadImage(String imageName, ImageType imageType){
         Path imagePath;
-        switch (imageType){
-            case AVATAR -> imagePath = Path.of(imageAvatarDirectory, imageName);
-            case ICON -> imagePath = Path.of(imageIconDirectory, imageName);
-            default -> imagePath = Path.of(imageDefaultDirectory, imageName);
+        switch (imageType) {
+            case AVATAR -> imagePath = Path.of(imageAvatarDirectory).resolve(imageName);
+            case ICON -> imagePath = Path.of(imageIconDirectory).resolve(imageName);
+            default -> imagePath = Path.of(imageDefaultDirectory).resolve(imageName);
         }
-
         try{
             if (Files.exists(imagePath)) {
                 return Files.readAllBytes(imagePath);
@@ -89,12 +110,19 @@ public class ImageService {
             throw new ImageLoadException();
         }
     }
-    public void deleteImage(String imageName, ImageType imageType){
+    public void deleteImage(Image image,ImageType imageType){
+        if(Boolean.TRUE.equals(image.getIs_default())){
+            return;
+        }
+        imageRepository.deleteByUuid(image.getUuid());
+        if(image.getFile()==null || image.getFile().isEmpty()){
+            return;
+        }
         Path imagePath;
-        switch (imageType){
-            case AVATAR -> imagePath = Path.of(imageAvatarDirectory, imageName);
-            case ICON -> imagePath = Path.of(imageIconDirectory, imageName);
-            default -> imagePath = Path.of(imageDefaultDirectory, imageName);
+        switch (imageType) {
+            case AVATAR -> imagePath = Path.of(imageAvatarDirectory, image.getFile());
+            case ICON -> imagePath = Path.of(imageIconDirectory, image.getFile());
+            default -> imagePath = Path.of(imageDefaultDirectory, image.getFile());
         }
         File fileToDelete = FileUtils.getFile(imagePath.toFile());
         FileUtils.deleteQuietly(fileToDelete);
@@ -115,5 +143,9 @@ public class ImageService {
             throw new IllegalStateException();
         }
         return multipartFile;
+    }
+    @Cacheable("defaultAvatar")
+    public Image getDefaultAvatar(){
+        return imageRepository.findByFile("guest.jpg").orElseThrow(ImageNullException::new);
     }
 }
